oc debug: Not deleting namespace if preserve-pod flag is passed
`--preserve-pod` flag in oc debug is used to prevent deletion of pod.
However, if temporary namespace is created to run this privileged pod,
oc debug removes this namespace in any case and that causes pod trying
to be prevented also garbage collected.

This PR prevents deletion of namespace if the `--preserve-pod` flag is
passed to keep debug pod in cluster.

+++ pkg/cli/debug/debug.go
@@ -1243,6 +1243,9 @@ func (o *DebugOptions) getNamespace(infoNs string) (string, func(), error) {
 		}
 
 		cleanup := func() {
+			if o.PreservePod {
+				return
+			}
 			if err := o.CoreClient.Namespaces().Delete(context.TODO(), ns.Name, metav1.DeleteOptions{}); err != nil {
 				klog.V(2).Infof("Unable to delete temporary namespace %s: %v", ns.Name, err)
 			} else {
